mktpub/requirements.txt
pyyaml
pika
requests
